Refactor to use collectors package, run collections in parallel
